#1302 - Allow auto-merging of incomplete but non-disagreeing annotations
- Migrate more tests into the test suite
