+++ webanno-curation/src/test/java/de/tudarmstadt/ukp/clarin/webanno/curation/casmerge/CasMergeTest.java
@@ -19,32 +19,18 @@ package de.tudarmstadt.ukp.clarin.webanno.curation.casmerge;
 
 import static de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.FEAT_REL_SOURCE;
 import static de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.FEAT_REL_TARGET;
-import static de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.RELATION_TYPE;
-import static de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.SPAN_TYPE;
 import static de.tudarmstadt.ukp.clarin.webanno.curation.CurationTestUtils.HOST_TYPE;
 import static de.tudarmstadt.ukp.clarin.webanno.curation.CurationTestUtils.createMultiLinkWithRoleTestTypeSytem;
-import static de.tudarmstadt.ukp.clarin.webanno.curation.CurationTestUtils.loadWebAnnoTSV;
-import static de.tudarmstadt.ukp.clarin.webanno.curation.CurationTestUtils.loadWebAnnoTsv3;
 import static de.tudarmstadt.ukp.clarin.webanno.curation.CurationTestUtils.makeLinkFS;
 import static de.tudarmstadt.ukp.clarin.webanno.curation.CurationTestUtils.makeLinkHostFS;
 import static de.tudarmstadt.ukp.clarin.webanno.curation.CurationTestUtils.makeLinkHostMultiSPanFeatureFS;
-import static de.tudarmstadt.ukp.clarin.webanno.curation.CurationTestUtils.readWebAnnoTSV;
 import static de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.LinkCompareBehavior.LINK_TARGET_AS_LABEL;
-import static de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.RelationDiffAdapter.DEPENDENCY_DIFF_ADAPTER;
-import static de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.SpanDiffAdapter.NER_DIFF_ADAPTER;
-import static de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.SpanDiffAdapter.POS_DIFF_ADAPTER;
 import static de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.doDiff;
-import static de.tudarmstadt.ukp.clarin.webanno.model.AnchoringMode.CHARACTERS;
-import static de.tudarmstadt.ukp.clarin.webanno.model.AnchoringMode.SINGLE_TOKEN;
 import static de.tudarmstadt.ukp.clarin.webanno.model.AnchoringMode.TOKENS;
 import static de.tudarmstadt.ukp.clarin.webanno.model.Mode.CURATION;
-import static de.tudarmstadt.ukp.clarin.webanno.model.OverlapMode.NO_OVERLAP;
-import static de.tudarmstadt.ukp.clarin.webanno.model.OverlapMode.OVERLAP_ONLY;
 import static java.util.Arrays.asList;
-import static org.apache.uima.fit.factory.AnalysisEngineFactory.createEngineDescription;
 import static org.apache.uima.fit.factory.JCasFactory.createJCas;
 import static org.apache.uima.fit.factory.JCasFactory.createText;
-import static org.apache.uima.fit.pipeline.SimplePipeline.runPipeline;
 import static org.apache.uima.fit.util.CasUtil.getType;
 import static org.apache.uima.fit.util.CasUtil.select;
 import static org.apache.uima.fit.util.CasUtil.selectCovered;
@@ -52,12 +38,7 @@ import static org.apache.uima.fit.util.JCasUtil.select;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assume.assumeTrue;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.when;
-import static org.mockito.MockitoAnnotations.initMocks;
 
-import java.io.File;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
@@ -65,7 +46,6 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
-import org.apache.commons.io.FileUtils;
 import org.apache.uima.cas.CAS;
 import org.apache.uima.cas.Feature;
 import org.apache.uima.cas.FeatureStructure;
@@ -75,398 +55,30 @@ import org.apache.uima.fit.factory.JCasFactory;
 import org.apache.uima.fit.util.CasUtil;
 import org.apache.uima.jcas.JCas;
 import org.assertj.core.api.Assertions;
-import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.mockito.Mock;
 
-import de.tudarmstadt.ukp.clarin.webanno.api.AnnotationSchemaService;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.exception.AnnotationException;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.feature.FeatureSupport;
-import de.tudarmstadt.ukp.clarin.webanno.api.annotation.feature.FeatureSupportRegistryImpl;
-import de.tudarmstadt.ukp.clarin.webanno.api.annotation.feature.PrimitiveUimaFeatureSupport;
-import de.tudarmstadt.ukp.clarin.webanno.api.annotation.feature.SlotFeatureSupport;
-import de.tudarmstadt.ukp.clarin.webanno.api.annotation.layer.ChainLayerSupport;
-import de.tudarmstadt.ukp.clarin.webanno.api.annotation.layer.LayerBehaviorRegistryImpl;
-import de.tudarmstadt.ukp.clarin.webanno.api.annotation.layer.LayerSupportRegistryImpl;
-import de.tudarmstadt.ukp.clarin.webanno.api.annotation.layer.RelationLayerSupport;
-import de.tudarmstadt.ukp.clarin.webanno.api.annotation.layer.SpanLayerSupport;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.model.AnnotatorState;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.model.AnnotatorStateImpl;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.util.WebAnnoCasUtil;
 import de.tudarmstadt.ukp.clarin.webanno.curation.CurationTestUtils;
 import de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff;
-import de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.DiffAdapter;
 import de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.DiffResult;
-import de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.RelationDiffAdapter;
 import de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.SpanDiffAdapter;
 import de.tudarmstadt.ukp.clarin.webanno.curation.casdiff.CasDiff.SpanPosition;
-import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationFeature;
-import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationLayer;
-import de.tudarmstadt.ukp.clarin.webanno.model.LinkMode;
-import de.tudarmstadt.ukp.clarin.webanno.model.MultiValueMode;
 import de.tudarmstadt.ukp.clarin.webanno.model.OverlapMode;
-import de.tudarmstadt.ukp.clarin.webanno.model.Project;
-import de.tudarmstadt.ukp.clarin.webanno.model.SourceDocument;
 import de.tudarmstadt.ukp.clarin.webanno.security.model.User;
-import de.tudarmstadt.ukp.clarin.webanno.tsv.WebannoTsv3XWriter;
 import de.tudarmstadt.ukp.dkpro.core.api.lexmorph.type.pos.POS;
-import de.tudarmstadt.ukp.dkpro.core.api.metadata.type.DocumentMetaData;
 import de.tudarmstadt.ukp.dkpro.core.api.ner.type.NamedEntity;
-import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Sentence;
 import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Token;
 import de.tudarmstadt.ukp.dkpro.core.api.syntax.type.dependency.Dependency;
 import de.tudarmstadt.ukp.dkpro.core.testing.DkproTestContext;
 
 public class CasMergeTest
+    extends CasMergeTestBase
 {
-    private @Mock AnnotationSchemaService schemaService;
-    
-    private CasMerge sut;
-    
-    private LayerSupportRegistryImpl layerSupportRegistry;
-    private FeatureSupportRegistryImpl featureSupportRegistry;
-    private Project project;
-    private AnnotationLayer sentenceLayer;
-    private AnnotationLayer tokenLayer;
-    private AnnotationFeature tokenPosFeature;
-    private AnnotationLayer posLayer;
-    private AnnotationFeature posFeature;
-    private AnnotationFeature posCoarseFeature;
-    private AnnotationLayer neLayer;
-    private AnnotationFeature neFeature;
-    private AnnotationFeature neIdentifierFeature;
-    private AnnotationLayer depLayer;
-    private AnnotationFeature depFeature;
-    private AnnotationFeature depFlavorFeature;
-    private AnnotationLayer slotLayer;
-    private AnnotationFeature slotFeature;
-    private AnnotationFeature stringFeature;
-    private AnnotationLayer multiValRel;
-    private AnnotationFeature multiValRelRel1;
-    private AnnotationFeature multiValRelRel2;
-    private AnnotationLayer multiValSpan;
-    private AnnotationFeature multiValSpanF1;
-    private AnnotationFeature multiValSpanF2;
-    private SourceDocument document;
-    private List<String> entryTypes;
-    
-    private static final RelationDiffAdapter MULTIVALREL_DIFF_ADAPTER = new RelationDiffAdapter(
-            "webanno.custom.Multivalrel", "Dependent", "Governor", "rel1", "rel2");
-    private static final SpanDiffAdapter MULTIVALSPAN_DIFF_ADAPTER = new SpanDiffAdapter(
-            "webanno.custom.Multivalspan", "f1", "f2");
-    
-    @Before
-    public void setup() throws Exception
-    {
-        initMocks(this);
-        
-        entryTypes = new ArrayList<>();
-        entryTypes.add(Sentence.class.getName());
-        entryTypes.add(Token.class.getName());
-        
-        project = new Project();
-        
-        document = new SourceDocument();
-        document.setProject(project);
-        document.setName("document");
-
-        sentenceLayer = new AnnotationLayer(Sentence.class.getName(), "Sentence", SPAN_TYPE, null,
-                true, CHARACTERS, NO_OVERLAP);
-
-        tokenLayer = new AnnotationLayer(Token.class.getName(), "Token", SPAN_TYPE, null, true,
-                CHARACTERS, NO_OVERLAP);
-        
-        tokenPosFeature = new AnnotationFeature();
-        tokenPosFeature.setName("pos");
-        tokenPosFeature.setEnabled(true);
-        tokenPosFeature.setType(POS.class.getName());
-        tokenPosFeature.setUiName("pos");
-        tokenPosFeature.setLayer(tokenLayer);
-        tokenPosFeature.setProject(project);
-        tokenPosFeature.setVisible(true);
-        
-        posLayer = new AnnotationLayer(POS.class.getName(), "POS", SPAN_TYPE, project, true,
-                SINGLE_TOKEN, NO_OVERLAP);
-        posLayer.setAttachType(tokenLayer);
-        posLayer.setAttachFeature(tokenPosFeature);
-        
-        posFeature = new AnnotationFeature();
-        posFeature.setName("PosValue");
-        posFeature.setEnabled(true);
-        posFeature.setType(CAS.TYPE_NAME_STRING);
-        posFeature.setUiName("PosValue");
-        posFeature.setLayer(posLayer);
-        posFeature.setProject(project);
-        posFeature.setVisible(true);
-
-        posCoarseFeature = new AnnotationFeature();
-        posCoarseFeature.setName("coarseValue");
-        posCoarseFeature.setEnabled(true);
-        posCoarseFeature.setType(CAS.TYPE_NAME_STRING);
-        posCoarseFeature.setUiName("coarseValue");
-        posCoarseFeature.setLayer(posLayer);
-        posCoarseFeature.setProject(project);
-        posCoarseFeature.setVisible(true);
-
-        neLayer = new AnnotationLayer(NamedEntity.class.getName(), "Named Entity", SPAN_TYPE,
-                project, true, TOKENS, OVERLAP_ONLY);
-        
-        neFeature = new AnnotationFeature();
-        neFeature.setName("value");
-        neFeature.setEnabled(true);
-        neFeature.setType(CAS.TYPE_NAME_STRING);
-        neFeature.setUiName("value");
-        neFeature.setLayer(neLayer);
-        neFeature.setProject(project);
-        neFeature.setVisible(true);
-
-        neIdentifierFeature = new AnnotationFeature();
-        neIdentifierFeature.setName("identifier");
-        neIdentifierFeature.setEnabled(true);
-        neIdentifierFeature.setType(CAS.TYPE_NAME_STRING);
-        neIdentifierFeature.setUiName("identifier");
-        neIdentifierFeature.setLayer(neLayer);
-        neIdentifierFeature.setProject(project);
-        neIdentifierFeature.setVisible(true);
-
-        depLayer = new AnnotationLayer(Dependency.class.getName(), "Dependency", RELATION_TYPE,
-                project, true, SINGLE_TOKEN, OVERLAP_ONLY);
-        depLayer.setAttachType(tokenLayer);
-        depLayer.setAttachFeature(tokenPosFeature);
-
-        depFeature = new AnnotationFeature();
-        depFeature.setName("DependencyType");
-        depFeature.setEnabled(true);
-        depFeature.setType(CAS.TYPE_NAME_STRING);
-        depFeature.setUiName("Relation");
-        depFeature.setLayer(depLayer);
-        depFeature.setProject(project);
-        depFeature.setVisible(true);
-
-        depFlavorFeature = new AnnotationFeature();
-        depFlavorFeature.setName("flavor");
-        depFlavorFeature.setEnabled(true);
-        depFlavorFeature.setType(CAS.TYPE_NAME_STRING);
-        depFlavorFeature.setUiName("flavor");
-        depFlavorFeature.setLayer(depLayer);
-        depFlavorFeature.setProject(project);
-        depFlavorFeature.setVisible(true);
-
-        slotLayer = new AnnotationLayer(HOST_TYPE, HOST_TYPE, SPAN_TYPE, project, false,
-                SINGLE_TOKEN, NO_OVERLAP);
-        
-        slotFeature = new AnnotationFeature();
-        slotFeature.setName("links");
-        slotFeature.setEnabled(true);
-        slotFeature.setType(Token.class.getName());
-        slotFeature.setMode(MultiValueMode.ARRAY);
-        slotFeature.setLinkMode(LinkMode.WITH_ROLE);
-        slotFeature.setLinkTypeName(CurationTestUtils.LINK_TYPE);
-        slotFeature.setLinkTypeRoleFeatureName("role");
-        slotFeature.setLinkTypeTargetFeatureName("target");
-        slotFeature.setUiName("links");
-        slotFeature.setLayer(slotLayer);
-        slotFeature.setProject(project);
-        slotFeature.setVisible(true);
-        
-        stringFeature = new AnnotationFeature();
-        stringFeature.setName("f1");
-        stringFeature.setEnabled(true);
-        stringFeature.setType(CAS.TYPE_NAME_STRING);
-        stringFeature.setUiName("f1");
-        stringFeature.setLayer(slotLayer);
-        stringFeature.setProject(project);
-        stringFeature.setVisible(true);
-        
-        multiValSpan = new AnnotationLayer("webanno.custom.Multivalspan", "Multivalspan", SPAN_TYPE,
-                project, true, TOKENS, OVERLAP_ONLY);
-        
-        multiValSpanF1 = new AnnotationFeature();
-        multiValSpanF1.setName("f1");
-        multiValSpanF1.setEnabled(true);
-        multiValSpanF1.setType(CAS.TYPE_NAME_STRING);
-        multiValSpanF1.setUiName("f1");
-        multiValSpanF1.setLayer(multiValSpan);
-        multiValSpanF1.setProject(project);
-        multiValSpanF1.setVisible(true);
-        
-        multiValSpanF2 = new AnnotationFeature();
-        multiValSpanF2.setName("f2");
-        multiValSpanF2.setEnabled(true);
-        multiValSpanF2.setType(CAS.TYPE_NAME_STRING);
-        multiValSpanF2.setUiName("f2");
-        multiValSpanF2.setLayer(multiValSpan);
-        multiValSpanF2.setProject(project);
-        multiValSpanF2.setVisible(true);
-        
-        multiValRel = new AnnotationLayer("webanno.custom.Multivalrel", "Multivalrel",
-                RELATION_TYPE, project, true, SINGLE_TOKEN, OVERLAP_ONLY);
-        multiValRel.setAttachType(multiValSpan);
-        
-        multiValRelRel1 = new AnnotationFeature();
-        multiValRelRel1.setName("rel1");
-        multiValRelRel1.setEnabled(true);
-        multiValRelRel1.setType(CAS.TYPE_NAME_STRING);
-        multiValRelRel1.setUiName("rel1");
-        multiValRelRel1.setLayer(multiValSpan);
-        multiValRelRel1.setProject(project);
-        multiValRelRel1.setVisible(true);
-        
-        multiValRelRel2 = new AnnotationFeature();
-        multiValRelRel2.setName("rel2");
-        multiValRelRel2.setEnabled(true);
-        multiValRelRel2.setType(CAS.TYPE_NAME_STRING);
-        multiValRelRel2.setUiName("rel2");
-        multiValRelRel2.setLayer(multiValSpan);
-        multiValRelRel2.setProject(project);
-        multiValRelRel2.setVisible(true);        
-        
-        when(schemaService.findLayer(any(Project.class), any(String.class))).thenAnswer(call -> {
-            String type = call.getArgument(1, String.class);
-            if (type.equals(Sentence.class.getName())) {
-                return sentenceLayer;
-            }
-            if (type.equals(Token.class.getName())) {
-                return tokenLayer;
-            }
-            if (type.equals(Dependency.class.getName())) {
-                return depLayer;
-            }
-            if (type.equals(POS.class.getName())) {
-                return posLayer;
-            }
-            if (type.equals(NamedEntity.class.getName())) {
-                return neLayer;
-            }
-            if (type.equals(CurationTestUtils.HOST_TYPE)) {
-                return slotLayer;
-            }
-            if (type.equals("webanno.custom.Multivalrel")) {
-                return multiValRel;
-            }
-            if (type.equals("webanno.custom.Multivalspan")) {
-                return multiValSpan;
-            }
-            throw new IllegalStateException("Unknown layer type: " + type);
-        });
-        
-        when(schemaService.listAnnotationFeature(any(AnnotationLayer.class))).thenAnswer(call -> { 
-            AnnotationLayer type = call.getArgument(0, AnnotationLayer.class);
-            if (type.getName().equals(Sentence.class.getName())) {
-                return asList();
-            }
-            if (type.getName().equals(Token.class.getName())) {
-                return asList();
-            }
-            if (type.getName().equals(Dependency.class.getName())) {
-                return asList(depFeature, depFlavorFeature);
-            }
-            if (type.getName().equals(POS.class.getName())) {
-                return asList(posFeature, posCoarseFeature);
-            }
-            if (type.getName().equals(NamedEntity.class.getName())) {
-                return asList(neFeature, neIdentifierFeature);
-            }
-            if (type.getName().equals(HOST_TYPE)) {
-                return asList(slotFeature, stringFeature);
-            }
-            if (type.getName().equals("webanno.custom.Multivalrel")) {
-                return asList(multiValRelRel1, multiValRelRel2);
-            }
-            if (type.getName().equals("webanno.custom.Multivalspan")) {
-                return asList(multiValSpanF1, multiValSpanF2);
-            }
-            throw new IllegalStateException("Unknown layer type: " + type.getName());
-        });
-
-        when(schemaService.getAdapter(any(AnnotationLayer.class))).thenAnswer(call -> { 
-            AnnotationLayer type = call.getArgument(0, AnnotationLayer.class);
-            return layerSupportRegistry.getLayerSupport(type).createAdapter(type);
-        });
-        
-        featureSupportRegistry = new FeatureSupportRegistryImpl(
-                asList(new PrimitiveUimaFeatureSupport(),
-                        new SlotFeatureSupport(schemaService)));
-        featureSupportRegistry.init();
-
-        LayerBehaviorRegistryImpl layerBehaviorRegistry = new LayerBehaviorRegistryImpl(asList());
-        layerBehaviorRegistry.init();
-
-        layerSupportRegistry = new LayerSupportRegistryImpl(asList(
-                new SpanLayerSupport(featureSupportRegistry, null, schemaService,
-                        layerBehaviorRegistry),
-                new RelationLayerSupport(featureSupportRegistry, null, schemaService,
-                        layerBehaviorRegistry),
-                new ChainLayerSupport(featureSupportRegistry, null, schemaService,
-                        layerBehaviorRegistry)));
-        layerSupportRegistry.init();
-        
-        sut = new CasMerge(schemaService);
-    }
-
-
-    @Test
-    public void simpleSpanNoDiffNoLabelTest()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = loadWebAnnoTSV(null, 
-                "mergecas/simplespan/1sentence.tsv",
-                "mergecas/simplespan/1sentence.tsv");
-        
-        JCas curatorCas = createText(casByUser.values().stream()
-                .flatMap(Collection::stream).findFirst().get().getDocumentText());
-
-        entryTypes.add(POS.class.getName());
-
-        List<SpanDiffAdapter> diffAdapters = asList(POS_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        CAS actual = readWebAnnoTSV("mergecas/simplespan/1sentence.tsv", null);
-        
-        casByUser = new HashMap<>();
-        casByUser.put("actual", asList(actual));
-        casByUser.put("merge", asList(curatorCas.getCas()));
-
-        result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        assertEquals(0, result.getDifferingConfigurationSets().size());
-        assertEquals(0, result.getIncompleteConfigurationSets().size());
-    }
-
-    @Test
-    public void simpleSpanDiffNoLabelTest()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = loadWebAnnoTSV(null, 
-                "mergecas/simplespan/1sentence.tsv",
-                "mergecas/simplespan/1sentenceempty.tsv");
-        
-        JCas curatorCas = createText(casByUser.values().stream()
-                .flatMap(Collection::stream).findFirst().get().getDocumentText());
-
-        entryTypes.add(POS.class.getName());
-
-        List<SpanDiffAdapter> diffAdapters = asList(POS_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        casByUser = new HashMap<>();
-        CAS actual = readWebAnnoTSV("mergecas/simplespan/1sentenceempty.tsv", null);
-        casByUser.put("actual", asList(actual));
-        casByUser.put("merge", asList(curatorCas.getCas()));
-
-        result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        assertEquals(0, result.getDifferingConfigurationSets().size());
-        assertEquals(0, result.getIncompleteConfigurationSets().size());
-    }
-
     /**
      * If one annotator has provided an annotation at a given position and the other annotator did
      * not (i.e. the annotations are incomplete), then this should be detected as a disagreement. 
@@ -475,8 +87,6 @@ public class CasMergeTest
     public void thatIncompleteAnnotationIsNotMerged()
         throws Exception
     {
-        entryTypes.add(POS.class.getName());
-        
         JCas user1 = JCasFactory.createText("word");
         token(user1, 0, 4, "X");
         
@@ -490,8 +100,7 @@ public class CasMergeTest
         JCas curatorCas = createText(casByUser.values().stream()
                 .flatMap(Collection::stream).findFirst().get().getDocumentText());
         
-        DiffResult result = doDiff(entryTypes, asList(POS_DIFF_ADAPTER), LINK_TARGET_AS_LABEL,
-                casByUser);
+        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
 
         sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
 
@@ -513,8 +122,6 @@ public class CasMergeTest
     public void thatIncompleteAnnotationIsMerged()
         throws Exception
     {
-        entryTypes.add(POS.class.getName());
-        
         JCas user1 = JCasFactory.createText("word");
         token(user1, 0, 4, "X");
         
@@ -528,8 +135,7 @@ public class CasMergeTest
         JCas curatorCas = createText(casByUser.values().stream()
                 .flatMap(Collection::stream).findFirst().get().getDocumentText());
         
-        DiffResult result = doDiff(entryTypes, asList(POS_DIFF_ADAPTER), LINK_TARGET_AS_LABEL,
-                casByUser);
+        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
 
         sut.setMergeIncompleteAnnotations(true);
         sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
@@ -559,423 +165,6 @@ public class CasMergeTest
     }
     
     @Test
-    public void simpleSpanNoDiffWithLabelTest()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = loadWebAnnoTSV(null,
-                "mergecas/simplespan/1sentenceposlabel.tsv",
-                "mergecas/simplespan/1sentenceposlabel.tsv");
-        
-        JCas curatorCas = createText(casByUser.values().stream()
-                .flatMap(Collection::stream).findFirst().get().getDocumentText());
-
-        entryTypes.add(POS.class.getName());
-
-        List<SpanDiffAdapter> diffAdapters = asList(POS_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        casByUser = new HashMap<>();
-        CAS actual = readWebAnnoTSV("mergecas/simplespan/1sentenceposlabel.tsv",
-                null);
-        casByUser.put("actual", asList(actual));
-        casByUser.put("merge", asList(curatorCas.getCas()));
-
-        result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        assertEquals(0, result.getDifferingConfigurationSets().size());
-        assertEquals(0, result.getIncompleteConfigurationSets().size());
-    }
-
-    @Test
-    public void simpleSpanDiffWithLabelTest()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = loadWebAnnoTSV(null,
-                "mergecas/simplespan/1sentenceposlabel.tsv",
-                "mergecas/simplespan/1sentenceposlabel2.tsv");
-        
-        JCas curatorCas = createText(casByUser.values().stream()
-                .flatMap(Collection::stream).findFirst().get().getDocumentText());
-
-        entryTypes.add(POS.class.getName());
-
-        List<SpanDiffAdapter> diffAdapters = asList(POS_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL,
-                casByUser);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        casByUser = new HashMap<>();
-        CAS actual = readWebAnnoTSV("mergecas/simplespan/1sentenceempty.tsv", null);
-        casByUser.put("actual", asList(actual));
-        casByUser.put("merge", asList(curatorCas.getCas()));
-
-        result = CasDiff.doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        assertEquals(0, result.getDifferingConfigurationSets().size());
-        assertEquals(0, result.getIncompleteConfigurationSets().size());
-    }
-
-    @Test
-    public void simpleSpanDiffIncompleteTest()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = loadWebAnnoTSV(null,
-                "mergecas/simplespan/1sentenceposlabel2.tsv",
-                "mergecas/simplespan/1sentenceposlabel3.tsv");
-        
-        JCas curatorCas = createText(casByUser.values().stream()
-                .flatMap(Collection::stream).findFirst().get().getDocumentText());
-
-        entryTypes.add(POS.class.getName());
-
-        List<SpanDiffAdapter> diffAdapters = asList(POS_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        CAS actual = readWebAnnoTSV("mergecas/simplespan/1sentenceposlabel2and3merged.tsv", null);
-        casByUser = new HashMap<>();
-        casByUser.put("actual", asList(actual));
-        casByUser.put("merge", asList(curatorCas.getCas()));
-
-        result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        assertEquals(0, result.getDifferingConfigurationSets().size());
-        assertEquals(2, result.getIncompleteConfigurationSets().size());
-    }
-    
-    @Test
-    public void simpleSpanDiffWithLabelStackingTest()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = loadWebAnnoTSV(null,
-                "mergecas/simplespan/1sentenceNEstacked.tsv",
-                "mergecas/simplespan/1sentenceNEstacked.tsv");
-        
-        JCas curatorCas = createText(casByUser.values().stream()
-                .flatMap(Collection::stream).findFirst().get().getDocumentText());
-
-        entryTypes.add(NamedEntity.class.getName());
-
-        List<SpanDiffAdapter> diffAdapters = asList(NER_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        CAS actual = readWebAnnoTSV("mergecas/simplespan/1sentenceNEempty.tsv", null);
-        casByUser = new HashMap<>();
-        casByUser.put("actual", asList(actual));
-        casByUser.put("merge", asList(curatorCas.getCas()));
-
-        result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        assertEquals(0, result.getDifferingConfigurationSets().size());
-        assertEquals(0, result.getIncompleteConfigurationSets().size());
-    }
-
-    @Test
-    public void simpleSpanDiffWithLabelStacking2Test()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = loadWebAnnoTSV(null,
-                "mergecas/simplespan/1sentenceNE.tsv",
-                "mergecas/simplespan/1sentenceNEstacked.tsv");
-        
-        JCas curatorCas = createText(casByUser.values().stream()
-                .flatMap(Collection::stream).findFirst().get().getDocumentText());
-
-        entryTypes.add(NamedEntity.class.getName());
-
-        List<SpanDiffAdapter> diffAdapters = asList(NER_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        CAS actual = readWebAnnoTSV("mergecas/simplespan/1sentenceNEempty.tsv", null);
-        casByUser = new HashMap<>();
-        casByUser.put("actual", asList(actual));
-        casByUser.put("merge", asList(curatorCas.getCas()));
-
-        result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        assertEquals(0, result.getDifferingConfigurationSets().size());
-        assertEquals(0, result.getIncompleteConfigurationSets().size());
-    }
-
-    @Test
-    public void thatMultipleStackedAnnotationsAreNotMerged()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = loadWebAnnoTSV(null,
-                // this document contains two stacked NE annoatations (ORG, ORGpart)
-                "mergecas/simplespan/1sentenceNE.tsv",
-                // this document contains three stacked NE annoatations (ORG, ORG, ORGpart)
-                "mergecas/simplespan/1sentenceNEstacked2.tsv");
-        
-        JCas curatorCas = createText(casByUser.values().stream()
-                .flatMap(Collection::stream).findFirst().get().getDocumentText());
-
-        entryTypes.add(NamedEntity.class.getName());
-
-        List<SpanDiffAdapter> diffAdapters = asList(NER_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        // If there are any stacked annotations, then no merging is performed
-        assertThat(select(curatorCas, NamedEntity.class))
-                .isEmpty();
-    }
-
-    @Test
-    public void simpleSpanNoDiffMultiFeatureTest()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = new HashMap<>();
-        casByUser.put("user1", asList(loadWebAnnoTsv3(
-                "testsuite/" + testContext.getMethodName() + "/user1.tsv")
-                        .getCas()));
-        casByUser.put("user2", asList(loadWebAnnoTsv3(
-                "testsuite/" + testContext.getMethodName() + "/user2.tsv")
-                        .getCas()));
-        JCas curatorCas = createText(casByUser.values().stream().flatMap(Collection::stream)
-                .findFirst().get().getDocumentText());
-
-        entryTypes.add(NamedEntity.class.getName());
-
-        List<? extends DiffAdapter> diffAdapters = asList(NER_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        result.print(System.out);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        assertMatch(curatorCas);
-    }
-
-    @Test
-    public void simpleSpanDiffMultiFeatureTest()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = new HashMap<>();
-        casByUser.put("user1", asList(loadWebAnnoTsv3(
-                "testsuite/" + testContext.getMethodName() + "/user1.tsv")
-                        .getCas()));
-        casByUser.put("user2", asList(loadWebAnnoTsv3(
-                "testsuite/" + testContext.getMethodName() + "/user2.tsv")
-                        .getCas()));
-        JCas curatorCas = createText(casByUser.values().stream().flatMap(Collection::stream)
-                .findFirst().get().getDocumentText());
-
-        entryTypes.add(NamedEntity.class.getName());
-
-        List<? extends DiffAdapter> diffAdapters = asList(NER_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        result.print(System.out);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        assertMatch(curatorCas);
-    }
-
-    @Test
-    public void simpleRelNoDiffTest()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = loadWebAnnoTSV(null,
-                "mergecas/rels/1sentencesamerel.tsv", 
-                "mergecas/rels/1sentencesamerel.tsv");
-        
-        JCas curatorCas = createText(casByUser.values().stream()
-                .flatMap(Collection::stream).findFirst().get().getDocumentText());
-
-        entryTypes.add(Dependency.class.getName());
-        entryTypes.add(POS.class.getName());
-
-        List<? extends DiffAdapter> diffAdapters = asList(
-                DEPENDENCY_DIFF_ADAPTER,
-                POS_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        CAS actual = CurationTestUtils.readWebAnnoTSV("mergecas/rels/1sentencesamerel.tsv", null);
-        casByUser = new HashMap<>();
-        casByUser.put("actual", asList(actual));
-        casByUser.put("merge", asList(curatorCas.getCas()));
-
-        result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        assertEquals(0, result.getDifferingConfigurationSets().size());
-        assertEquals(0, result.getIncompleteConfigurationSets().size());
-    }
-
-    @Test
-    public void simpleRelGovDiffTest()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = loadWebAnnoTSV(null,
-                "mergecas/rels/1sentencesamerel.tsv", 
-                "mergecas/rels/1sentencesamerel2.tsv");
-        
-        JCas curatorCas = createText(casByUser.values().stream()
-                .flatMap(Collection::stream).findFirst().get().getDocumentText());
-
-        entryTypes.add(Dependency.class.getName());
-        entryTypes.add(POS.class.getName());
-
-        List<? extends DiffAdapter> diffAdapters = asList(
-                DEPENDENCY_DIFF_ADAPTER,
-                POS_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        casByUser = new HashMap<>();
-        CAS actual = readWebAnnoTSV("mergecas/rels/1sentencesamerel3.tsv", null);
-        casByUser.put("actual", asList(actual));
-        casByUser.put("merge", asList(curatorCas.getCas()));
-
-        result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        assertEquals(0, result.getDifferingConfigurationSets().size());
-        assertEquals(0, result.getIncompleteConfigurationSets().size());
-    }
-
-    @Test
-    public void simpleRelTypeDiffTest()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = loadWebAnnoTSV(null,
-                "mergecas/rels/1sentencesamerel.tsv", 
-                "mergecas/rels/1sentencesamerel4.tsv");
-        
-        JCas curatorCas = createText(casByUser.values().stream()
-                .flatMap(Collection::stream).findFirst().get().getDocumentText());
-
-        entryTypes.add(Dependency.class.getName());
-        entryTypes.add(POS.class.getName());
-
-        List<? extends DiffAdapter> diffAdapters = asList(
-                DEPENDENCY_DIFF_ADAPTER,
-                POS_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        CAS actual = readWebAnnoTSV("mergecas/rels/1sentencesamerel5.tsv", null);
-        casByUser = new HashMap<>();
-        casByUser.put("actual", asList(actual));
-        casByUser.put("merge", asList(curatorCas.getCas()));
-
-        result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        assertEquals(0, result.getDifferingConfigurationSets().size());
-        assertEquals(0, result.getIncompleteConfigurationSets().size());
-    }
-
-    @Test
-    public void simpleRelGovStackedTest()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = new HashMap<>();
-        casByUser.put("user1", asList(loadWebAnnoTsv3(
-                "testsuite/" + testContext.getMethodName() + "/user1.tsv")
-                        .getCas()));
-        casByUser.put("user2", asList(loadWebAnnoTsv3(
-                "testsuite/" + testContext.getMethodName() + "/user2.tsv")
-                        .getCas()));
-        JCas curatorCas = createText(casByUser.values().stream().flatMap(Collection::stream)
-                .findFirst().get().getDocumentText());
-
-        entryTypes.add("webanno.custom.Multivalrel");
-        entryTypes.add("webanno.custom.Multivalspan");
-
-        List<? extends DiffAdapter> diffAdapters = asList(MULTIVALREL_DIFF_ADAPTER,
-                MULTIVALSPAN_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        result.print(System.out);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        assertMatch(curatorCas);        
-    }
-
-    @Test
-    public void relStackedTest()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = new HashMap<>();
-        casByUser.put("user1", asList(loadWebAnnoTsv3(
-                "testsuite/" + testContext.getMethodName() + "/user1.tsv")
-                        .getCas()));
-        casByUser.put("user2", asList(loadWebAnnoTsv3(
-                "testsuite/" + testContext.getMethodName() + "/user2.tsv")
-                        .getCas()));
-        JCas curatorCas = createText(casByUser.values().stream().flatMap(Collection::stream)
-                .findFirst().get().getDocumentText());
-
-        entryTypes.add("webanno.custom.Multivalrel");
-        entryTypes.add("webanno.custom.Multivalspan");
-
-        List<? extends DiffAdapter> diffAdapters = asList(MULTIVALREL_DIFF_ADAPTER,
-                MULTIVALSPAN_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        result.print(System.out);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        assertMatch(curatorCas);        
-    }
-
-    @Test
-    public void relationLabelTest()
-        throws Exception
-    {
-        Map<String, List<CAS>> casByUser = new HashMap<>();
-        casByUser.put("user1", asList(loadWebAnnoTsv3(
-                "testsuite/" + testContext.getMethodName() + "/user1.tsv")
-                        .getCas()));
-        casByUser.put("user2", asList(loadWebAnnoTsv3(
-                "testsuite/" + testContext.getMethodName() + "/user2.tsv")
-                        .getCas()));
-        JCas curatorCas = createText(casByUser.values().stream().flatMap(Collection::stream)
-                .findFirst().get().getDocumentText());
-
-        entryTypes.add(POS.class.getName());
-        entryTypes.add(Dependency.class.getName());
-
-        List<? extends DiffAdapter> diffAdapters = asList(DEPENDENCY_DIFF_ADAPTER);
-
-        DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
-
-        result.print(System.out);
-
-        sut.reMergeCas(result, document, null, curatorCas.getCas(), getSingleCasByUser(casByUser));
-
-        assertMatch(curatorCas);
-    }
-
-    @Test
     public void multiLinkWithRoleNoDifferenceTest()
         throws Exception
     {
@@ -994,12 +183,6 @@ public class CasMergeTest
         JCas curatorCas = createJCas(createMultiLinkWithRoleTestTypeSytem("f1"));
         curatorCas.setDocumentText(casByUser.values().stream().flatMap(Collection::stream)
                 .findFirst().get().getDocumentText());
-
-        entryTypes.add(HOST_TYPE);
-
-        SpanDiffAdapter adapter = new SpanDiffAdapter(HOST_TYPE);
-        adapter.addLinkFeature("links", "role", "target");
-        List<? extends DiffAdapter> diffAdapters = asList(adapter);
 
         DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
 
@@ -1035,12 +218,6 @@ public class CasMergeTest
         curatorCas.setDocumentText(casByUser.values().stream().flatMap(Collection::stream)
                 .findFirst().get().getDocumentText());
 
-        entryTypes.add(HOST_TYPE);
-
-        SpanDiffAdapter adapter = new SpanDiffAdapter(HOST_TYPE);
-        adapter.addLinkFeature("links", "role", "target");
-        List<? extends DiffAdapter> diffAdapters = asList(adapter);
-
         DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
 
         result.print(System.out);
@@ -1075,12 +252,6 @@ public class CasMergeTest
         JCas curatorCas = createJCas(createMultiLinkWithRoleTestTypeSytem("f1"));
         curatorCas.setDocumentText(casByUser.values().stream().flatMap(Collection::stream)
                 .findFirst().get().getDocumentText());
-
-        entryTypes.add(HOST_TYPE);
-
-        SpanDiffAdapter adapter = new SpanDiffAdapter(HOST_TYPE);
-        adapter.addLinkFeature("links", "role", "target");
-        List<? extends DiffAdapter> diffAdapters = asList(adapter);
 
         DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
 
@@ -1119,8 +290,6 @@ public class CasMergeTest
         curatorCas.setDocumentText(casByUser.values().stream().flatMap(Collection::stream)
                 .findFirst().get().getDocumentText());
 
-        entryTypes.add(HOST_TYPE);
-
         SpanDiffAdapter adapter = new SpanDiffAdapter(HOST_TYPE);
         adapter.addLinkFeature("links", "role", "target");
         
@@ -1156,12 +325,6 @@ public class CasMergeTest
         JCas curatorCas = createJCas(createMultiLinkWithRoleTestTypeSytem("f1"));
         curatorCas.setDocumentText(casByUser.values().stream().flatMap(Collection::stream)
                 .findFirst().get().getDocumentText());
-
-        entryTypes.add(HOST_TYPE);
-
-        SpanDiffAdapter adapter = new SpanDiffAdapter(HOST_TYPE);
-        adapter.addLinkFeature("links", "role", "target");
-        List<? extends DiffAdapter> diffAdapters = asList(adapter);
 
         DiffResult result = doDiff(entryTypes, diffAdapters, LINK_TARGET_AS_LABEL, casByUser);
 
@@ -1370,12 +533,6 @@ public class CasMergeTest
         casByUser.put("user1", asList(mergeCas.getCas()));
         casByUser.put("user2", asList(jcasA.getCas()));
 
-        List<String> entryTypes = asList(CurationTestUtils.HOST_TYPE);
-
-        CasDiff.SpanDiffAdapter adapter = new CasDiff.SpanDiffAdapter(CurationTestUtils.HOST_TYPE);
-        adapter.addLinkFeature("links", "role", "target");
-        List<? extends CasDiff.DiffAdapter> diffAdapters = asList(adapter);
-
         CasDiff.DiffResult diff = CasDiff.doDiff(entryTypes, diffAdapters,
                 CasDiff.LinkCompareBehavior.LINK_TARGET_AS_LABEL, casByUser);
 
@@ -1409,12 +566,6 @@ public class CasMergeTest
         Map<String, List<CAS>> casByUser = new LinkedHashMap<>();
         casByUser.put("user1", asList(mergeCas.getCas()));
         casByUser.put("user2", asList(jcasA.getCas()));
-
-        List<String> entryTypes = asList(CurationTestUtils.HOST_TYPE);
-
-        CasDiff.SpanDiffAdapter adapter = new CasDiff.SpanDiffAdapter(CurationTestUtils.HOST_TYPE);
-        adapter.addLinkFeature("links", "role", "target");
-        List<? extends CasDiff.DiffAdapter> diffAdapters = asList(adapter);
 
         CasDiff.DiffResult diff = CasDiff.doDiff(entryTypes, diffAdapters,
                 CasDiff.LinkCompareBehavior.LINK_TARGET_AS_LABEL, casByUser);
@@ -1591,28 +742,28 @@ public class CasMergeTest
                 .withMessageContaining("annotation already exists");
     }
     
-    private void assertMatch(JCas curatorCas)
-        throws Exception
-    {
-        String referenceFolder = "src/test/resources/testsuite/" + testContext.getMethodName();
-        File targetFolder = testContext.getTestOutputFolder();
-        
-        DocumentMetaData dmd = DocumentMetaData.get(curatorCas);
-        dmd.setDocumentId("curator");
-        runPipeline(curatorCas, createEngineDescription(WebannoTsv3XWriter.class,
-                WebannoTsv3XWriter.PARAM_TARGET_LOCATION, targetFolder,
-                WebannoTsv3XWriter.PARAM_OVERWRITE, true));
-        
-        File referenceFile = new File(referenceFolder, "curator.tsv");
-        assumeTrue("No reference data available for this test.", referenceFile.exists());
-        
-        File actualFile = new File(targetFolder, "curator.tsv");
-        
-        String reference = FileUtils.readFileToString(referenceFile, "UTF-8");
-        String actual = FileUtils.readFileToString(actualFile, "UTF-8");
-        
-        assertEquals(reference, actual);
-    }
+//    private void writeTestSuiteData(Map<String, List<CAS>> casByUser, JCas curatorCas)
+//        throws Exception
+//    {
+//        runPipeline(casByUser.get("user1").get(0),
+//                createEngineDescription(WebannoTsv3XWriter.class,
+//                        WebannoTsv3XWriter.PARAM_SINGULAR_TARGET, true,
+//                        WebannoTsv3XWriter.PARAM_OVERWRITE, true,
+//                        WebannoTsv3XWriter.PARAM_TARGET_LOCATION,
+//                        "target/bux/" + testContext.getMethodName() + "/user1.tsv"));
+//        runPipeline(casByUser.get("user2").get(0),
+//                createEngineDescription(WebannoTsv3XWriter.class,
+//                        WebannoTsv3XWriter.PARAM_SINGULAR_TARGET, true,
+//                        WebannoTsv3XWriter.PARAM_OVERWRITE, true,
+//                        WebannoTsv3XWriter.PARAM_TARGET_LOCATION,
+//                        "target/bux/" + testContext.getMethodName() + "/user2.tsv"));
+//        runPipeline(curatorCas,
+//                createEngineDescription(WebannoTsv3XWriter.class,
+//                        WebannoTsv3XWriter.PARAM_SINGULAR_TARGET, true,
+//                        WebannoTsv3XWriter.PARAM_OVERWRITE, true,
+//                        WebannoTsv3XWriter.PARAM_TARGET_LOCATION,
+//                        "target/bux/" + testContext.getMethodName() + "/curator.tsv"));    
+//    }
     
     @Rule
     public DkproTestContext testContext = new DkproTestContext();
