+++ webanno-curation/src/test/resources/testsuite/simpleRelNoDiffTest/user1.tsv
@@ -1,0 +1,15 @@
+#FORMAT=WebAnno TSV 3.2
+#T_SP=de.tudarmstadt.ukp.dkpro.core.api.lexmorph.type.pos.POS|PosValue|coarseValue
+#T_RL=de.tudarmstadt.ukp.dkpro.core.api.syntax.type.dependency.Dependency|DependencyType|flavor|BT_de.tudarmstadt.ukp.dkpro.core.api.lexmorph.type.pos.POS
+
+
+#Text=This is a simple sentence to be executed .
+1-1	0-4	This	_	_	_	_	_	
+1-2	5-7	is	VAFIN	*	_	_	_	
+1-3	8-9	a	_	_	_	_	_	
+1-4	10-16	simple	_	_	_	_	_	
+1-5	17-25	sentence	NN	*	OBJA	*	1-2	
+1-6	26-28	to	_	_	_	_	_	
+1-7	29-31	be	_	_	_	_	_	
+1-8	32-40	executed	_	_	_	_	_	
+1-9	41-42	.	_	_	_	_	_
