+++ webanno-curation/src/test/resources/testsuite/simpleSpanDiffIncompleteTest/user1.tsv
@@ -1,0 +1,14 @@
+#FORMAT=WebAnno TSV 3.2
+#T_SP=de.tudarmstadt.ukp.dkpro.core.api.lexmorph.type.pos.POS|PosValue|coarseValue
+
+
+#Text=This is a simple sentence to be executed . 
+1-1	0-4	This	_	_	
+1-2	5-7	is	_	_	
+1-3	8-9	a	_	_	
+1-4	10-16	simple	_	_	
+1-5	17-25	sentence	NE	*	
+1-6	26-28	to	_	_	
+1-7	29-31	be	_	_	
+1-8	32-40	executed	_	_	
+1-9	41-42	.	_	_
